float to int II

+++ TPO_1C2017/src/tda/DMMedicionesTDA.java
@@ -5,7 +5,7 @@ public interface DMMedicionesTDA {
 
 	public void inicializar();
 	
-	public void agregar(int anio, int mes, int dia, float medicion);
+	public void agregar(int anio, int mes, int dia, int medicion);
 	
 	public void eliminarAnio(int anio);
